Changed 'tick' variables to 64 bit
- This fixes an issue with timers that stop working after about 24-49
  days when the tick overflows (note that this may happen much earlier
  than that, and at hard to predict times, on some systems)
- Updated the RDTSC help message in the configure script to also warn
  users about issues with SpeedStep enabled systems.
- On Windows, tick() still has a resolution of 10~15ms (or even as low
  as 100ms on some systems). A TODO comment (thanks, Ai4rei) was added
  for a follow-up patch, as I want this one to be as small as possible)
- Note: on Windows versions earlier than 6.x (Vista, Server 2008), the
  tick overflow issue is NOT fixed, since they don't support the
  function used to retrieve a 64 bit tick. This isn't a big issue, since
  those platforms are already - or going soon to be - out of their
  extended support period, and it's already advisable to upgrade, for
  other reasons. If you're the unfortunate user of such a system, it is
  recommended that you reboot your machine at least once every 49 days
  for Hercules to work reliably.
- Note: To clear some doubts, since I've already been asked, this has
  absolutely NOTHING to do with 32/64 bit CPUs or OSes. It's all about
  a variable's size, not the size of registers of your CPU, and your
  32bit CPU will be able to handle this just fine.

+++ src/login/ipban_sql.c
@@ -36,7 +36,7 @@ static Sql* sql_handle = NULL;
 static int cleanup_timer_id = INVALID_TIMER;
 static bool ipban_inited = false;
 
-int ipban_cleanup(int tid, unsigned int tick, int id, intptr_t data);
+int ipban_cleanup(int tid, int64 tick, int id, intptr_t data);
 
 
 // initialize
@@ -247,8 +247,7 @@ void ipban_log(uint32 ip)
 }
 
 // remove expired bans
-int ipban_cleanup(int tid, unsigned int tick, int id, intptr_t data)
-{
+int ipban_cleanup(int tid, int64 tick, int id, intptr_t data) {
 	if( !login_config.ipban )
 		return 0;// ipban disabled
